chore: move remaining Prometheus API helpers into internal/prometheus
